requirements.txt
flask==3.1.1
flask-cors==6.0.1
requests==2.32.4
beautifulsoup4==4.13.4
python-dotenv==1.0.0
